Z3r0D4y-By-ZER0L0G0N-2025/paytm_karo/requirements.txt
pillow==12.0.0
pyzbar==0.1.9
